Some refactorings.
Some comments were added with the idea of making the code more
understandable.

And an icon was added in an attempt to change the way the app looks.

And, the app only updates the movie list when the sort criteria
preference has changed. Before, it was updated each time the main
activity was started; it didn't matter if the user changed the sorting
preferences or not.